Continue work on BL-175 docs
