+++ src/tab_foundry/bench/iris.py
@@ -1,9 +1,13 @@
-"""Iris evaluation helpers for tab-foundry checkpoints."""
+"""Iris evaluation helpers and CLI for tab-foundry checkpoints."""
 
 from __future__ import annotations
 
+import argparse
 from dataclasses import dataclass
 from pathlib import Path
+from typing import Sequence
+
+import numpy as np
 
 from sklearn.datasets import load_iris
 from sklearn.ensemble import RandomForestClassifier
@@ -91,3 +95,31 @@ def evaluate_iris_checkpoint(
             )
         )
     return IrisEvalSummary(checkpoint=checkpoint_path.expanduser().resolve(), results=results)
+
+
+def build_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(description="Evaluate a tab-foundry checkpoint on binary Iris")
+    parser.add_argument("--checkpoint", required=True, help="Classification checkpoint path")
+    parser.add_argument(
+        "--device",
+        default="cpu",
+        choices=("cpu", "cuda", "mps"),
+        help="Inference device for the checkpointed model",
+    )
+    parser.add_argument("--seeds", type=int, default=5, help="Number of train/test splits")
+    return parser
+
+
+def main(argv: Sequence[str] | None = None) -> int:
+    args = build_parser().parse_args(argv)
+    summary = evaluate_iris_checkpoint(
+        Path(str(args.checkpoint)),
+        device=str(args.device),
+        seeds=int(args.seeds),
+    )
+    print(f"Iris evaluation for checkpoint={summary.checkpoint}")
+    print(f"{'Method':<15} {'ROC AUC':>8} {'Std':>8}")
+    print("-" * 33)
+    for name, values in sorted(summary.results.items(), key=lambda item: -np.mean(item[1])):
+        print(f"{name:<15} {np.mean(values):>8.3f} {np.std(values):>8.3f}")
+    return 0
